runtime: fix rent calculation bug

+++ src/flamenco/runtime/sysvar/fd_sysvar_rent.c
@@ -53,7 +53,7 @@ ulong
 fd_rent_exempt_minimum_balance2( fd_rent_t const * rent,
                                  ulong             data_len ) {
   /* https://github.com/solana-labs/solana/blob/792fafe0c25ac06868e3ac80a2b13f1a5b4a1ef8/sdk/program/src/rent.rs#L72 */
-  return (data_len + ACCOUNT_STORAGE_OVERHEAD) * ((ulong) ((double)rent->lamports_per_uint8_year * rent->exemption_threshold));
+  return (ulong)( (double)((data_len + ACCOUNT_STORAGE_OVERHEAD) * rent->lamports_per_uint8_year) * (double)rent->exemption_threshold );
 }
 
 ulong fd_rent_exempt_minimum_balance( fd_global_ctx_t* global, ulong data_len ) {
